requirements.txt
Django>=4.2.0
batchalign>=0.1.0
pyannote.audio>=3.1.1
torch>=2.0.0
torchaudio>=2.0.0
pydub>=0.25.1
numpy>=1.24.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.0
requests>=2.31.0
speechbrain>=0.5.14
scikit-learn>=1.3.0 